requirement.txt
numpy
pandas
tensorflow-cpu
scikit-learn
yfinance
fastapi
uvicorn
aiohttp
python-multipart
scipy
jinja2